fix(scheduler-service): export event attendee dto and renamed it
SFO-108
