+++ services/scheduler-service/src/models/event-attendee-view-item.dto.ts
